Move printData back into SecurityLogsFragment
